Extract shared helpers into pyriemann/_helpers.py
Introduce pyriemann/_helpers.py as a leaf module containing the
validation, type-check, and decorator utilities used by both
pyriemann.geometry and pyriemann.utils. It has zero internal pyriemann
imports — only stdlib (warnings) and numpy / array_api_compat.

Helpers extracted: deprecated, check_function, check_init, check_like,
check_matrix_pair, check_weights, is_real_type, is_square.

Source modules in pyriemann/utils/ now thin-import these helpers from
_helpers and re-export them, so all prior public names at
pyriemann.utils.utils, pyriemann.utils.test, pyriemann.utils.docs and
pyriemann.utils.deprecated continue to resolve unchanged.

Inside pyriemann/geometry/, the modules now import the helpers directly
from pyriemann._helpers instead of going through pyriemann.utils.{utils,
test,docs}. The only remaining utils dependency is pyriemann.utils._backend
(the array-API backend selector), which is itself a leaf.
