+++ pyriemann/geometry/covariance.py
@@ -5,7 +5,14 @@ from array_api_compat import array_namespace as get_namespace, device as xpd
 from array_api_extra import expand_dims
 from scipy.stats import chi2
 
-from pyriemann.utils.docs import deprecated
+from pyriemann._helpers import (
+    check_function,
+    check_init,
+    check_weights,
+    deprecated,
+    is_real_type,
+    is_square,
+)
 from pyriemann.utils._backend import (
     apply_xp_cov,
     diag_indices,
@@ -16,8 +23,6 @@ from pyriemann.utils._backend import (
 )
 from .base import ctranspose, _vectorize_nd
 from .distance import distance_mahalanobis
-from pyriemann.utils.test import is_square, is_real_type
-from pyriemann.utils.utils import check_function, check_init, check_weights
 
 
 def _cov(X, **kwds):
