+++ pyriemann/geometry/distance.py
@@ -7,10 +7,9 @@ from array_api_compat import (
 )
 from sklearn.metrics import euclidean_distances
 
+from pyriemann._helpers import check_function, check_matrix_pair, is_real_type
 from pyriemann.utils._backend import diag_indices, torch, tril_indices
 from .base import _eigvalsh, invsqrtm, logm, powm, sqrtm
-from pyriemann.utils.test import is_real_type
-from pyriemann.utils.utils import check_function, check_matrix_pair
 
 
 ###############################################################################
