+++ pyriemann/geometry/tangentspace.py
@@ -8,9 +8,9 @@ from array_api_compat import (
     is_numpy_namespace,
 )
 
+from pyriemann._helpers import check_function, check_matrix_pair
 from pyriemann.utils._backend import diag_indices, tril_indices, triu_indices
 from .base import ctranspose, expm, invsqrtm, logm, sqrtm, ddexpm, ddlogm
-from pyriemann.utils.utils import check_function, check_matrix_pair
 
 
 def exp_map_euclid(X, Cref):
