+++ pyriemann/utils/utils.py
@@ -1,117 +1,12 @@
 import inspect
 
-import array_api_compat.numpy as _xpnp
-from array_api_compat import array_namespace as get_namespace, device as xpd
-
-
-def check_like(like):
-    """Resolve array-API namespace and device from a reference array.
-
-    Parameters
-    ----------
-    like : None | ndarray
-        Reference array. If None, returns the array-API NumPy namespace
-        and a ``None`` device.
-
-    Returns
-    -------
-    xp : module
-        The array-API namespace (NumPy or PyTorch).
-    dev : object | None
-        The device of ``like`` if provided, else ``None``.
-
-    Notes
-    -----
-    .. versionadded:: 0.12
-    """
-    if like is None:
-        return _xpnp, None
-    return get_namespace(like), xpd(like)
-
-
-def check_matrix_pair(A, B, *, require_square=False):
-    """Validate two matrix arrays share compatible matrix dimensions.
-
-    Parameters
-    ----------
-    A, B : ndarray, shape (..., n, m)
-        Input matrix arrays.
-    require_square : bool, default=False
-        If True, also require the matrix dimensions to be square.
-
-    Returns
-    -------
-    xp : module
-        The shared array-API namespace of ``A`` and ``B``.
-
-    Notes
-    -----
-    .. versionadded:: 0.12
-    """
-    xp = get_namespace(A, B)
-    if A.ndim < 2 or B.ndim < 2:
-        raise ValueError("Inputs must be at least 2D arrays")
-    if A.shape[-2:] != B.shape[-2:]:
-        raise ValueError("Inputs must have equal matrix dimensions")
-    if require_square and A.shape[-2] != A.shape[-1]:
-        raise ValueError("Inputs must contain square matrices")
-    if A.shape != B.shape:
-        try:
-            xp.broadcast_shapes(A.shape[:-2], B.shape[:-2])
-        except (ValueError, RuntimeError) as exc:
-            raise ValueError("Inputs have incompatible dimensions.") from exc
-    return xp
-
-
-def check_weights(weights, n_weights, *, check_positivity=False, like=None):
-    """Check weights.
-
-    If input is None, output weights are equal.
-    Strict positivity of weights can be checked.
-    In any case, weights are normalized (sum equal to 1).
-
-    Parameters
-    ----------
-    weights : None | ndarray, shape (n_weights,), default=None
-        Input weights. If None, it provides equal weights.
-    n_weights : int
-        Number of weights to provide if None, or to check.
-    check_positivity : bool, default=False
-        Choose if strict positivity of weights is checked.
-    like : None | ndarray, default=None
-        Reference array used to infer the array-API namespace and device of
-        the returned weights. If None, NumPy is used.
-
-        .. versionadded:: 0.12
-
-    Returns
-    -------
-    weights : ndarray, shape (n_weights,)
-        Output checked weights.
-
-    Notes
-    -----
-    .. versionadded:: 0.4
-    .. versionchanged:: 0.12
-        Add support for NumPy and PyTorch.
-    """
-    xp, dev = check_like(like)
-
-    if weights is None:
-        weights = xp.ones(n_weights, dtype=float, device=dev)
-
-    else:
-        weights = xp.asarray(weights, device=dev)
-        if weights.shape != (n_weights,):
-            raise ValueError(
-                "Weights do not have the good shape. Should be (%d,) but got "
-                "%s." % (n_weights, weights.shape,)
-            )
-        if check_positivity and bool(xp.any(weights <= 0)):
-            raise ValueError("Weights must be strictly positive.")
-
-    weights = weights / xp.sum(weights)
-    return weights
+from pyriemann._helpers import (  # noqa: F401
+    check_function,
+    check_init,
+    check_like,
+    check_matrix_pair,
+    check_weights,
+)
 
 
 def check_metric(metric, expected_keys=["mean", "distance"]):
@@ -150,78 +45,6 @@ def check_metric(metric, expected_keys=["mean", "distance"]):
 
     else:
         raise TypeError(f"metric must be str or dict, but got {type(metric)}")
-
-
-def check_function(fun, functions):
-    """Check the function to use.
-
-    Parameters
-    ----------
-    fun : string | callable
-        Function to check.
-        If string, it must be one of the keys of ``functions``.
-        If callable, it can be a function defined in API or by the user.
-        In the latter case, the signature of the function as to match the ones
-        defined in ``functions``. This is the user responsibility to ensure
-        this, and will not be checked.
-    functions : dict
-        Functions available in API, used only when ``fun`` is a string.
-
-    Returns
-    -------
-    fun : callable
-        Function to use.
-
-    Notes
-    -----
-    .. versionadded:: 0.6
-    """
-    if isinstance(fun, str):
-        if fun not in functions.keys():
-            raise ValueError(f"Unknown function name '{fun}'. Must be one of "
-                             f"{' '.join(functions.keys())}")
-        else:
-            fun = functions[fun]
-    elif not hasattr(fun, '__call__'):
-        raise ValueError("Argument must be a string or a callable "
-                         f"(Got {type(fun)}).")
-    return fun
-
-
-def check_init(init, n, *, like=None):
-    """Check the initial matrix.
-
-    Parameters
-    ----------
-    init : ndarray, shape (n, n)
-        A square matrix used to initialize the algorithm.
-    n : int
-        Expected dimension of the matrix.
-    like : None | ndarray, default=None
-        Reference array used to infer the array-API namespace and device of
-        the returned matrix. If None, NumPy is used.
-
-        .. versionadded:: 0.12
-
-    Returns
-    -------
-    init : ndarray, shape (n, n)
-        The checked square matrix used to initialize the algorithm.
-
-    Notes
-    -----
-    .. versionadded:: 0.8
-    .. versionchanged:: 0.12
-        Add support for NumPy and PyTorch.
-    """
-    xp, dev = check_like(like)
-    init = xp.asarray(init, dtype=init.dtype, device=dev)
-    if init.shape != (n, n):
-        raise ValueError(
-            "Init matrix does not have the good shape. "
-            f"Should be ({n},{n}) but got {init.shape}."
-        )
-    return init
 
 
 def check_param_in_func(param, func):
